requirements.txt
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
biopython==1.78
bleach==3.3.0
boto3==1.18.46
botocore==1.21.46
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==5.0.5
defusedxml==0.7.1
entrypoints==0.3
fair-esm==0.4.0
filelock==3.0.12
idna==2.10
immutables==0.15
importlib-metadata==3.10.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.18.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
tape-proteins==0.5
tensorboardX==2.4
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
